move permissions to separate component
