+++ src/app/users/user-profile.component.html
@@ -163,36 +163,6 @@
         Permissions
       </mat-panel-title>
     </mat-expansion-panel-header>
-    <div class="user-profile-settings-section">
-      <p class="user-profile-settings-explanation">All users listed in the following field will be able to view your
-        relationships to skills including the corresponding levels and priorities.</p>
-      <mat-form-field class="user-profile-form-field">
-        <mat-chip-list #authorizedUsersChipList>
-          <mat-chip *ngFor="let authorizedUser of authorizedUsers"
-                    selectable="true" removable="true"
-                    (removed)="onAuthorizedUserRemoved(authorizedUser)">
-            {{ authorizedUser.firstName }} {{ authorizedUser.lastName }} ({{ authorizedUser.userName }})
-            <mat-icon matChipRemove>cancel</mat-icon>
-          </mat-chip>
-          <input placeholder="Users authorized to view my skills" #authorizedUsersInput
-                 [formControl]="authorizedUsersControl"
-                 [matAutocomplete]="authorizedUsersAutocomplete"
-                 [matChipInputFor]="authorizedUsersChipList"
-                 matChipInputAddOnBlur="true">
-        </mat-chip-list>
-        <mat-autocomplete #authorizedUsersAutocomplete="matAutocomplete"
-                          (optionSelected)="onAuthorizedUserSuggestionSelected($event)">
-          <mat-option *ngFor="let authorizedUser of authorizedUserSuggestions$ | async" [value]="authorizedUser">
-            {{ authorizedUser.firstName }} {{ authorizedUser.lastName }} ({{ authorizedUser.userName }})
-          </mat-option>
-        </mat-autocomplete>
-      </mat-form-field>
-    </div>
-    <mat-divider></mat-divider>
-    <div class="user-profile-actions">
-      <button mat-raised-button color="primary" [disabled]="savingInProgress" (click)="savePermissions()">Save
-        permissions
-      </button>
-    </div>
+    <app-permissions [savingInProgress]="savingInProgress"></app-permissions>
   </mat-expansion-panel>
 </div>
