+++ src/app/users/user-profile.component.spec.ts
@@ -1,5 +1,5 @@
-import { async, ComponentFixture, discardPeriodicTasks, fakeAsync, TestBed, tick } from '@angular/core/testing';
-import { Component, DebugElement } from '@angular/core';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { Component, DebugElement, Input } from '@angular/core';
 import { By } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { LayoutModule } from '@angular/cdk/layout';
@@ -45,6 +45,14 @@ class PublicationsStubComponent {
 class MembershipsStubComponent {
 }
 
+@Component({
+  selector: 'app-permissions',
+  template: ''
+})
+class PermissionsStubComponent {
+  @Input() savingInProgress = false;
+}
+
 describe('UserProfileComponent', () => {
   let component: UserProfileComponent;
   let fixture: ComponentFixture<UserProfileComponent>;
@@ -69,18 +77,6 @@ describe('UserProfileComponent', () => {
         }
       ));
 
-    spyOn(usersServiceStub, 'getAuthorizedUsers')
-      .and.returnValue(of<User[]>(
-        [{
-          id: '2736a204-f3ab-4b65-8568-a1c8ce1db8ab',
-          userName: 'testing',
-          firstName: 'Tina',
-          lastName: 'Testing',
-          email: 'tina.testing@skoop.io',
-          coach: false,
-        }]
-      ));
-
     spyOn(usersServiceStub, 'updateUser')
       .and.returnValue(of<User>(
         {
@@ -93,26 +89,6 @@ describe('UserProfileComponent', () => {
         }
       ));
 
-    spyOn(usersServiceStub, 'updateAuthorizedUsers')
-      .and.returnValue(of<User[]>(
-        [{
-          id: '2736a204-f3ab-4b65-8568-a1c8ce1db8ab',
-          userName: 'testing',
-          firstName: 'Tina',
-          lastName: 'Testing',
-          email: 'tina.testing@skoop.io',
-          coach: false,
-        },
-        {
-          id: '251c2a3b-b737-4622-8060-196d5e297ebc',
-          userName: 'testbed',
-          firstName: 'Tabia',
-          lastName: 'Testbed',
-          email: 'tabia.testbed@skoop.io',
-          coach: false,
-        }]
-      ));
-
     TestBed.configureTestingModule({
       imports: [
         BrowserAnimationsModule,
@@ -121,7 +97,7 @@ describe('UserProfileComponent', () => {
         ReactiveFormsModule,
         AppMaterialModule
       ],
-      declarations: [UserProfileComponent, TestimonialsStubComponent, PublicationsStubComponent, MembershipsStubComponent],
+      declarations: [UserProfileComponent, TestimonialsStubComponent, PublicationsStubComponent, MembershipsStubComponent, PermissionsStubComponent],
       providers: [
         GlobalErrorHandlerService,
         { provide: UsersService, useValue: usersServiceStub },
@@ -160,45 +136,6 @@ describe('UserProfileComponent', () => {
     expect(component.userForm.get('coach').value).toBeFalsy();
   });
 
-  it('should initialize the list of authorized users', () => {
-    expect(component.authorizedUsers).toEqual([{
-      id: '2736a204-f3ab-4b65-8568-a1c8ce1db8ab',
-      userName: 'testing',
-      firstName: 'Tina',
-      lastName: 'Testing',
-      email: 'tina.testing@skoop.io',
-      coach: false,
-    }]);
-  });
-
-  it('should remove a given user from the list of authorized users', () => {
-    component.authorizedUsers = [{
-      id: '2736a204-f3ab-4b65-8568-a1c8ce1db8ab',
-      userName: 'testing',
-      firstName: 'Tina',
-      lastName: 'Testing',
-      email: 'tina.testing@skoop.io',
-      coach: false,
-    },
-    {
-      id: '251c2a3b-b737-4622-8060-196d5e297ebc',
-      userName: 'testbed',
-      firstName: 'Tabia',
-      lastName: 'Testbed',
-      email: 'tabia.testbed@skoop.io',
-      coach: false,
-    }];
-    component.onAuthorizedUserRemoved(component.authorizedUsers[0]);
-    expect(component.authorizedUsers).toEqual([{
-      id: '251c2a3b-b737-4622-8060-196d5e297ebc',
-      userName: 'testbed',
-      firstName: 'Tabia',
-      lastName: 'Testbed',
-      email: 'tabia.testbed@skoop.io',
-      coach: false,
-    }]);
-  });
-
   it('should update the user profile form', async(() => {
     component.userForm.get('coach').setValue(true);
     component.saveUserDetails();
@@ -224,74 +161,6 @@ describe('UserProfileComponent', () => {
 
       expect(usersServiceStub.updateUser).toHaveBeenCalledWith(expectedRequestData);
     });
-  }));
-
-  it('should update the list of authorized users', async(() => {
-    component.authorizedUsers.push({
-      id: '251c2a3b-b737-4622-8060-196d5e297ebc',
-      userName: 'testbed',
-      firstName: 'Tabia',
-      lastName: 'Testbed',
-      email: 'tabia.testbed@skoop.io',
-      coach: false,
-    });
-    component.savePermissions();
-    fixture.whenStable().then(() => {
-      fixture.detectChanges();
-      expect(component.authorizedUsers).toEqual([{
-        id: '2736a204-f3ab-4b65-8568-a1c8ce1db8ab',
-        userName: 'testing',
-        firstName: 'Tina',
-        lastName: 'Testing',
-        email: 'tina.testing@skoop.io',
-        coach: false,
-      },
-      {
-        id: '251c2a3b-b737-4622-8060-196d5e297ebc',
-        userName: 'testbed',
-        firstName: 'Tabia',
-        lastName: 'Testbed',
-        email: 'tabia.testbed@skoop.io',
-        coach: false,
-      }]);
-
-      expect(usersServiceStub.updateAuthorizedUsers).toHaveBeenCalledWith(
-        UserPermissionScope.READ_USER_SKILLS,
-        [{
-          id: '2736a204-f3ab-4b65-8568-a1c8ce1db8ab',
-          userName: 'testing',
-          firstName: 'Tina',
-          lastName: 'Testing',
-          email: 'tina.testing@skoop.io',
-          coach: false,
-        },
-        {
-          id: '251c2a3b-b737-4622-8060-196d5e297ebc',
-          userName: 'testbed',
-          firstName: 'Tabia',
-          lastName: 'Testbed',
-          email: 'tabia.testbed@skoop.io',
-          coach: false,
-        }]
-      );
-    });
-  }));
-
-  it('should send getUserSuggestions request in 500 ms', fakeAsync(() => {
-    const usersService = TestBed.get(UsersService) as UsersService;
-    spyOn(usersService, 'getUserSuggestions').and.returnValue(of([]));
-
-    component.authorizedUsersControl.setValue('test');
-    tick(200);
-    fixture.detectChanges();
-
-    expect(usersService.getUserSuggestions).not.toHaveBeenCalled();
-
-    tick(300);
-    fixture.detectChanges();
-    expect(usersService.getUserSuggestions).toHaveBeenCalled();
-
-    discardPeriodicTasks();
   }));
 
   it('should add new elem to the language array', () => {
